extracted deck adapter into it's own class to abstract visualisation of the list away from the activity
